replaced Tooltipster lib with Magento tooltips
removed unnecessary template rewrite
added proper JS mixin
