+++ view/frontend/layout/catalog_product_view.xml
@@ -1,10 +1,6 @@
 <?xml version="1.0"?>
 <page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
       xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
-    <head>
-        <css src="Spaggel_Tooltip::css/main.css" />
-        <css src="Spaggel_Tooltip::css/tooltipster.bundle.min.css" />
-    </head>
     <body>
         <referenceBlock name="product.attributes" template="Spaggel_Tooltip::product/view/attributes.phtml"/>
     </body>
