--- view/frontend/templates/product/view/renderer.phtml
@@ -1,35 +1,0 @@
-<?php
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
-// @codingStandardsIgnoreFile
-//"swatchRenderer": {
-?>
-<?php /** @var $block \Magento\Swatches\Block\Product\Renderer\Configurable */ ?>
-<div class="swatch-opt" data-role="swatch-options"></div>
-
-<script type="text/x-magento-init">
-    {
-        "[data-role=swatch-options]": {
-            "SwatchRenderer": {
-                "jsonConfig": <?= /* @escapeNotVerified */ $swatchOptions = $block->getJsonConfig() ?>,
-                "jsonSwatchConfig": <?php /* @escapeNotVerified */
-                    echo $swatchOptions = $block->getJsonSwatchConfig(); ?>,
-                "mediaCallback": "<?= /* @escapeNotVerified */ $block->getMediaCallback() ?>",
-                "gallerySwitchStrategy": "<?php /* @escapeNotVerified */ echo $block->getVar('gallery_switch_strategy',
-                    'Magento_ConfigurableProduct') ?: 'replace'; ?>"
-            }
-        }
-    }
-</script>
-<script type="text/javascript">
-    require(['jquery', 'Tooltipster'], function($, Tooltipster) {
-        // init. tooltipster after the swatched are loaded because they're loaded with AJAX
-        $('[data-role=swatch-options]').on('swatch.initialized', function () {
-            new Tooltipster({
-                "theme": "tooltipster-custom"
-            }, $('.swatch-tooltip'));
-        });
-    })
-</script>
