--- view/frontend/web/css/main.css
@@ -1,97 +1,0 @@
-/* Tooltip icon */
-.tooltip-content-wrapper {
-    display: none;
-}
-.tooltip, .swatch-tooltip {
-    margin-left: 8px;
-    border-radius: 50%;
-    width: 20px;
-    height: 20px;
-    display: inline-block;
-    text-align: center;
-    color: #fff;
-    background: #116fb2;
-    text-decoration: none;
-}
-.tooltip:hover, .swatch-tooltip:hover {
-    background: #5da3ff;
-    color: #fff;
-    cursor: pointer;
-}
-
-/* THEME */
-.tooltipster-sidetip.tooltipster-custom .tooltipster-box {
-    border-radius: 3px;
-    border: 1px solid #cccccc;
-    background: #ededed;
-}
-
-.tooltipster-sidetip.tooltipster-custom .tooltipster-content {
-    color: #666666;
-}
-
-.tooltipster-sidetip.tooltipster-custom .tooltipster-arrow {
-    height: 9px;
-    margin-left: -9px;
-    width: 18px;
-}
-
-.tooltipster-sidetip.tooltipster-custom.tooltipster-left .tooltipster-arrow,
-.tooltipster-sidetip.tooltipster-custom.tooltipster-right .tooltipster-arrow {
-    height: 18px;
-    margin-left: 0;
-    margin-top: -9px;
-    width: 9px;
-}
-
-.tooltipster-sidetip.tooltipster-custom .tooltipster-arrow-background {
-    border: 9px solid transparent;
-}
-
-.tooltipster-sidetip.tooltipster-custom.tooltipster-bottom .tooltipster-arrow-background {
-    border-bottom-color: #ededed;
-    top: 1px;
-}
-
-.tooltipster-sidetip.tooltipster-custom.tooltipster-left .tooltipster-arrow-background {
-    border-left-color: #ededed;
-    left: -1px;
-}
-
-.tooltipster-sidetip.tooltipster-custom.tooltipster-right .tooltipster-arrow-background {
-    border-right-color: #ededed;
-    left: 1px;
-}
-
-.tooltipster-sidetip.tooltipster-custom.tooltipster-top .tooltipster-arrow-background {
-    border-top-color: #ededed;
-    top: -1px;
-}
-
-.tooltipster-sidetip.tooltipster-custom .tooltipster-arrow-border {
-    border: 9px solid transparent;
-}
-
-.tooltipster-sidetip.tooltipster-custom.tooltipster-bottom .tooltipster-arrow-border {
-    border-bottom-color: #cccccc;
-}
-
-.tooltipster-sidetip.tooltipster-custom.tooltipster-left .tooltipster-arrow-border {
-    border-left-color: #cccccc;
-}
-
-.tooltipster-sidetip.tooltipster-custom.tooltipster-right .tooltipster-arrow-border {
-    border-right-color: #cccccc;
-}
-
-.tooltipster-sidetip.tooltipster-custom.tooltipster-top .tooltipster-arrow-border {
-    border-top-color: #cccccc;
-}
-
-.tooltipster-sidetip.tooltipster-custom.tooltipster-bottom .tooltipster-arrow-uncropped {
-    top: -9px;
-}
-
-.tooltipster-sidetip.tooltipster-custom.tooltipster-right .tooltipster-arrow-uncropped {
-    left: -9px;
-}
