+++ view/frontend/web/css/source/_module.less
@@ -1,0 +1,3 @@
+.spaggel-tooltip {
+  .lib-tooltip(right);
+}
