--- view/frontend/web/js/tooltip.wrapper.js
@@ -1,10 +1,0 @@
-define([
-    'jquery',
-    'Spaggel_Tooltip/js/tooltipster.bundle.min'
-],function($){
-    return function (config, element) {
-        $(document).ready(function () {
-            $(element).tooltipster(config)
-        });
-    }
-});
